refactor: split all tasks into separate files
