+++ include/ams.h
@@ -1,0 +1,9 @@
+
+#ifndef AMS_H
+#define AMS_H
+
+#define AMS_TASK_PERIOD (10)
+
+void startAMSTask(void* pvParameters);
+
+#endif // AMS_H
