+++ include/torque.h
@@ -1,0 +1,9 @@
+
+#ifndef TORQUE_H
+#define TORQUE_H
+
+#define TORQUE_TASK_PERIOD (3)
+
+void startTorqueTask(void* pvParameters);
+
+#endif // TORQUE_H
